Renamed LTC68xx driver files in preparation for 6811->6813 changes
ltc6811_btm.c -> ltc6813_btm.c
ltc6811_btm.h -> ltc6813_btm.h
ltc6811_btm_bal.c -> ltc6813_btm_bal.c
ltc6811_btm_bal.h -> ltc6813_btm_bal.h
